Removed debugging code and fixed tests.
Various ``p`` statements that were used for debugging purposes have been
removed. I've also fixed the broken specifications by defining
`Innate.middleware_spec`.

Looking at this requirement I'm not entirely sure if having to define a new
method for a custom mode is the way to go. While it certainly keeps things
simple it should be documented *very clearly* if we stick with this.

Mentioning the following people for feedback on this: @manveru, @leucos.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>